hsearch-feature-examples: Fix mapstruct configuration to use jakarta.inject instead of javax.inject
For some reason when doing just 'mvn clean install' all was fine:
mapstruct was picking Jakarta and ignoring the (incorrect) config;
but not in dev mode.

Maybe because with 'mvn clean install' javax.inject isn't in the
compile classpath, whereas when Quarkus dev mode triggers compilation
it is?

In any case, the previous config was definitely invalid.